fix #3 (#6)
Changed Transaction Amount type from Long to BigDecimal

+++ calculation/src/main/java/ee/qrental/calculation/application/service/strategy/RateCalculationStrategy.java
@@ -2,8 +2,10 @@ package ee.qrental.calculation.application.service.strategy;
 
 import ee.qrental.calculation.application.service.QWeek;
 
+import java.math.BigDecimal;
+
 public interface RateCalculationStrategy {
     boolean canApply(final QWeek week);
 
-    Long calculate(final QWeek week);
+    BigDecimal calculate(final QWeek week);
 }
